+++ workspaces/adoption-insights/plugins/analytics-module-adoption-insights/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @red-hat-developer-hub/backstage-plugin-analytics-module-adoption-insights
 
+## 0.9.1
+
+### Patch Changes
+
+- d623d36: Refreshed `yarn.lock` to align resolved dependency versions.
+
 ## 0.9.0
 
 ### Minor Changes